api/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.1.0
django-filter==23.3
drf-spectacular==0.26.5
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
